Convert HOURS_UNTIL_TRASH_PERMANENTLY_DELETED to int before using it

Closes #1499

See merge request bramw/baserow!1210

+++ backend/src/baserow/config/settings/base.py
@@ -557,8 +557,8 @@ MAX_FIELD_LIMIT = 1500
 
 # If you change this default please also update the default for the web-frontend found
 # in web-frontend/modules/core/module.js:55
-HOURS_UNTIL_TRASH_PERMANENTLY_DELETED = os.getenv(
-    "HOURS_UNTIL_TRASH_PERMANENTLY_DELETED", 24 * 3
+HOURS_UNTIL_TRASH_PERMANENTLY_DELETED = int(
+    os.getenv("HOURS_UNTIL_TRASH_PERMANENTLY_DELETED", 24 * 3)
 )
 OLD_TRASH_CLEANUP_CHECK_INTERVAL_MINUTES = 5
